requirements.txt
anthropic>=0.44.0
fastmcp
fast-agent-mcp>=0.2.5
slack-bolt>=1.18.0
slack-sdk>=3.38.0
nextcord>=2.4.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
mcp-server-fetch
google-genai>=0.2.0
pillow>=10.0.0
PyYAML>=6.0.2
pydantic>=2.9.2
